Create a helper class for FDeltaUp and FDeltaDown
The functions for FDeltaUp and FDeltaDown are nearly identical. The
only difference between the two is which image is subtracted from the
other (in other words, whether the resulting image will be positive or
negative). Since the the rest of the code for the two classes is
identical, it seems prudent to create a helper class for both FDeltaUp
and down to rely on. By specifying to Delta (the helper class) whether
it is being called by FDeltaUp or FDeltaDown, it affects the order of
the image subtraction.