WIP: add tests for Delta F
Unfortunately, the Delta F function is really broken.
So broken that I don't have time to fix it.

* It destroys the original input image.
* It seems like it might timing issues.
* It probably doesn't work right with hyperstacks.
* It doesn't work headless.

This last point in particular makes it very hard to unit test.

So this branch is probably going to sit forever, unless an interested
person picks up the torch and completes the work... probably by
rewriting the Delta F function completely.