Move all cookbook plugins to fiji.cookbook package
The unnamed package causes too many problems.

Today, for example, there was a clash between the Cell Counter plugin
added to Fiji, and the one that ships with the Cookbook. But the
problems won't stop there. Let's just give the cookbook a package like
it should have had from the beginning.