Add missing plugin dependencies
See the following forum thread for discussion:
    http://forum.imagej.net/t/nucleus-counter-in-cookbook/2802

The Cookbook plugins here are an adapted subset of the ones from
the old MBF ImageJ collection (http://imagej.net/MBF). It turns out
that some plugins necessary for certain functionality were not
migrated over to the Cookbook repository, because they were not
explicit compile-time references of the calling plugins. This is
because IJ.run invokes plugins by name, not class, so there is no
type safety: if the plugin is missing, we don't know until runtime.

To prevent future problems, I did a thorough check of all places
where IJ.run was used to call a plugin. Here is the complete list:

    $ git grep -h 'IJ.run("' | sed 's/.*IJ.run(\("[^",]*"\).*/\1/' | sort -u
    "8-bit"
    "Adapative3DThreshold "
    "Add..."
    "Analyze Particles..."
    "Convert to Mask"
    "Duplicate..."
    "Entropy Threshold"
    "Fill"
    "Grays"
    "Image Calculator..."
    "Invert"
    "Mean..."
    "Median..."
    "Mixture Modeling threshold"
    "Multi Measure"
    "Multiply..."
    "OtsuThresholding 16Bit"
    "OtsuThresholding 8Bit"
    "RGB Color"
    "Red"
    "Rename..."
    "Scale Bar..."
    "Spectrum"
    "Subtract Background..."
    "Watershed"
    "k-means Clustering"

Of the above commands, the following are not part of IJ1 core:

    "Adapative3DThreshold "
    "Entropy Threshold"
    "Mixture Modeling threshold"
    "Multi Measure"
    "OtsuThresholding 16Bit"
    "OtsuThresholding 8Bit"
    "k-means Clustering"

Of those, the following were easily extractable from the
old MBF ImageJ plugins archive:

    "Adapative3DThreshold "
    "Entropy Threshold"
    "OtsuThresholding 16Bit"
    "OtsuThresholding 8Bit"