Add LICENSE file
Legally, the combined Fiji Cookbook work must be licensed under the GPL,
because some of the code here is GPL. (Try "git grep GNU" to find it.)

To make this clearer, let's add a LICENSE file declaring that
the Fiji Cookbook as a whole is a GPL'ed work.

Normally, I would also run "mvn license:update-file-header" to
prepend the GPL license header blurb to each and every file.
However, the current list of copyright authors according to the
POM is quite misleading, since these plugins originally came from
all over. Rather than deal with potentially doing it incorrectly,
we leave it unspecified, at least for now.